restructure bool_index module
